requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
websockets>=12.0
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
